# pkg/handler/middleware.go
package handler

import (
	"errors"
	"github.com/gin-gonic/gin"
	"github.com/sirupsen/logrus"
	"net/http"
	"strings"
)

const (
	authorizationHeader = "Authorization"
	userCtx             = "userId"
)

// @Summary Identity
// @Security ApiKeyAuth
// @Tags auth
// @Description authentication
// @ID authentication
// @Accept json
// @Produce json
// @Success 200
// @Failure 401 {object} errorResponse
// @Router /auth/identity [post]
func (h *Handler) userIdentity(ctx *gin.Context) {
	header := ctx.GetHeader(authorizationHeader)
	if header == "" {
		newErrorResponse(ctx, http.StatusUnauthorized, "auth header is empty")
		return
	}

	headerParts := strings.Split(header, " ")
	if len(headerParts) != 2 {
		newErrorResponse(ctx, http.StatusUnauthorized, "invalid auth header")
		return
	}

	claims, err := h.services.Authorization.ParseAccessToken(headerParts[1])
	if err != nil {
		newErrorResponse(ctx, http.StatusUnauthorized, err.Error())
		return
	}

	if _, err := h.services.GetSessionById(claims.SessionId); err != nil {
		logrus.WithFields(logrus.Fields{
			"package":  "handler",
			"file":     "middleware.go",
			"function": "userIdentity",
			"message":  err,
		}).Errorf("error getting session")
		newErrorResponse(ctx, http.StatusUnauthorized, err.Error())
		return
	}
}

func getUserId(ctx *gin.Context) (int, error) {
	id, ok := ctx.Get(userCtx)
	if !ok {
		newErrorResponse(ctx, http.StatusInternalServerError, "user id not found")
		return 0, errors.New("user id not found")
	}

	idInt, ok := id.(int)
	if !ok {
		newErrorResponse(ctx, http.StatusInternalServerError, "user id not found")
		return 0, errors.New("user id not found")
	}

	return idInt, nil
}